To add a new column with the cumulative dose. To add multiple columns use addDrugUtilisation() for efficiency.

To add a new column with the cumulative dose. To add multiple columns use addDrugUtilisation() for efficiency.

Usage

addCumulativeDose(
  cohort,
  ingredientConceptId,
  conceptSet = NULL,
  indexDate = "cohort_start_date",
  censorDate = "cohort_end_date",
  restrictIncident = TRUE,
  nameStyle = "cumulative_dose_{concept_name}_{ingredient}",
  name = NULL
)

Arguments

cohort

Cohort in the cdm.

ingredientConceptId

Ingredient OMOP concept that we are interested for the study. It is a compulsory input, no default value is provided.

conceptSet

List of concepts to be included. If NULL all the descendants of ingredient concept id will be used.

indexDate

Name of a column that indicates the date to start the analysis.

censorDate

Name of a column that indicates the date to stop the analysis, if NULL end of individuals observation is used.

restrictIncident

Whether to include only incident prescriptions in the analysis. If FALSE all prescriptions that overlap with the study period will be included.

nameStyle

Character string to specify the nameStyle of the new columns.

name

Name of the new computed column if NULL a temporary tables is created.

Value

The same cohort with the added column.

Examples

# \donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

cdm$cohort1|>
  addCumulativeDose(ingredientConceptId = 1125315)
#> # Source:   table<og_008_1720179961> [10 x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1022-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1          1 2020-09-20        2021-03-18     
#>  2                    3          2 2022-05-21        2022-06-05     
#>  3                    1          7 2015-09-10        2018-05-02     
#>  4                    3          8 2012-02-04        2012-02-15     
#>  5                    2          3 2010-02-10        2010-07-21     
#>  6                    1          4 2022-01-26        2022-04-28     
#>  7                    3          5 2019-07-17        2019-12-26     
#>  8                    3          6 1991-05-25        1992-04-18     
#>  9                    2          9 2022-10-05        2022-11-12     
#> 10                    1         10 1997-11-07        2003-08-04     
#> # ℹ 1 more variable:
#> #   cumulative_dose_ingredient_1125315_descendants_1125315 <dbl>
# }