To add a new column with the exposed time. To add multiple columns use addDrugUtilisation() for efficiency.

To add a new column with the exposed time. To add multiple columns use addDrugUtilisation() for efficiency.

Usage

addExposedTime(
  cohort,
  conceptSet,
  indexDate = "cohort_start_date",
  censorDate = "cohort_end_date",
  restrictIncident = TRUE,
  gapEra = 1,
  nameStyle = "exposed_time_{concept_name}",
  name = NULL
)

Arguments

cohort

Cohort in the cdm.

conceptSet

List of concepts to be included.

indexDate

Name of a column that indicates the date to start the analysis.

censorDate

Name of a column that indicates the date to stop the analysis, if NULL end of individuals observation is used.

restrictIncident

Whether to include only incident prescriptions in the analysis. If FALSE all prescriptions that overlap with the study period will be included.

gapEra

Number of days between two continuous exposures to be considered in the same era.

nameStyle

Character string to specify the nameStyle of the new columns.

name

Name of the new computed cohort table, if NULL a temporary tables is created.

Value

The same cohort with the added column.

Examples

# \donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()
codelist <- getDrugIngredientCodes(cdm, name = "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(
  cdm = cdm, name = "dus_cohort", conceptSet = codelist
)

cdm$dus_cohort |>
  addExposedTime(conceptSet = codelist)
#> # Source:   table<og_048_1721669169> [?? x 5]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1023-azure:R 4.4.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                   <int>      <int> <date>            <date>         
#>  1                    1          3 2009-04-18        2009-11-07     
#>  2                    1          6 1991-01-28        1991-09-17     
#>  3                    1          7 2018-05-21        2018-07-20     
#>  4                    1          1 2021-02-25        2021-03-11     
#>  5                    1          2 2022-05-24        2022-06-02     
#>  6                    1          9 2022-09-11        2022-11-09     
#>  7                    1          4 2021-05-05        2021-12-15     
#>  8                    1          2 2022-06-06        2022-06-07     
#>  9                    1          5 2014-07-28        2019-11-20     
#> 10                    1          7 2016-02-01        2017-10-04     
#> # ℹ more rows
#> # ℹ 1 more variable: exposed_time_161_acetaminophen <int>
# }