To add a new column with the duration of the first exposure. To add multiple columns use addDrugUtilisation() for efficiency.

To add a new column with the duration of the first exposure. To add multiple columns use addDrugUtilisation() for efficiency.

Usage

addInitialExposureDuration(
  cohort,
  conceptSet,
  indexDate = "cohort_start_date",
  censorDate = "cohort_end_date",
  restrictIncident = TRUE,
  nameStyle = "initial_exposure_duration_{concept_name}",
  name = NULL
)

Arguments

cohort

A cohort_table object.

conceptSet

List of concepts to be included.

indexDate

Name of a column that indicates the date to start the analysis.

censorDate

Name of a column that indicates the date to stop the analysis, if NULL end of individuals observation is used.

restrictIncident

Whether to include only incident prescriptions in the analysis. If FALSE all prescriptions that overlap with the study period will be included.

nameStyle

Character string to specify the nameStyle of the new columns.

name

Name of the new computed cohort table, if NULL a temporary table will be created.

Value

The same cohort with the added column.

Examples

# \donttest{
library(DrugUtilisation)
library(CodelistGenerator)
library(dplyr, warn.conflicts = FALSE)

cdm <- mockDrugUtilisation()
codelist <- getDrugIngredientCodes(cdm = cdm, name = "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(cdm = cdm,
                                        name = "dus_cohort",
                                        conceptSet = codelist)
#> ℹ Subsetting drug_exposure table
#> ℹ Checking whether any record needs to be dropped.
#> ℹ Collapsing overlaping records.
#> ℹ Collapsing records with gapEra = 1 days.

cdm$dus_cohort |>
  addInitialExposureDuration(conceptSet = codelist) |>
  glimpse()
#> Rows: ??
#> Columns: 5
#> Database: DuckDB v1.2.2 [unknown@Linux 6.11.0-1014-azure:R 4.5.0/:memory:]
#> $ cohort_definition_id                        <int> 1, 1, 1, 1, 1, 1, 1
#> $ subject_id                                  <int> 1, 8, 10, 10, 3, 4, 5
#> $ cohort_start_date                           <date> 2020-02-29, 2015-07-10, 20…
#> $ cohort_end_date                             <date> 2020-02-29, 2015-11-25, 20…
#> $ initial_exposure_duration_161_acetaminophen <int> 1, 139, 175, 7, 909, 105,…
# }