Run benchmark of drug utilisation cohort generation

Run benchmark of drug utilisation cohort generation

Usage

benchmarkDUS(
  cdm,
  numberOfCohort = 1:4,
  indicationCohortName = "cohort1",
  ingredientId = 1125315,
  drugExposureName = "drug_exposure"
)

Arguments

cdm

A CDM reference object

numberOfCohort

Number of cohort to generate for benchmarking. An integer or a vector of integers

indicationCohortName

Name of indication cohort table

ingredientId

Ingredient OMOP concept that we are interested for the study. It is a compulsory input, no default value is provided.

drugExposureName

Name of drug_exposure table in cdm, the table must contain drug_concept_id, quantity, drug_exposure_start_date and drug_exposure_end_date as columns

Value

a tibble with time taken for different analyses

Examples

# \donttest{
cdm <- DrugUtilisation::mockDrugUtilisation()
timings <- DrugUtilisation::benchmarkDUS(cdm)
#> Warning: The `durationRange` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.6.2.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `imputeDuration` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.6.2.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `priorObservation` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.6.2.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `cohortDateRange` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.6.2.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `limit` argument of `generateDrugUtilisationCohortSet()` is deprecated as
#> of DrugUtilisation 0.6.2.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Washout specified is 0 meaning nothing will be applied. Suggest rechecking the
#> priorUseWashout.
#> Warning: The `cdm` argument of `summariseDrugUse()` is deprecated as of DrugUtilisation
#> 0.5.0.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-09 08:40:36.459081
#> ✔ Summary finished, at 2024-07-09 08:40:36.501604
#> Washout specified is 0 meaning nothing will be applied. Suggest rechecking the
#> priorUseWashout.
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-09 08:40:53.80514
#> ✔ Summary finished, at 2024-07-09 08:40:53.842233
#> Washout specified is 0 meaning nothing will be applied. Suggest rechecking the
#> priorUseWashout.
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-09 08:41:11.234152
#> ✔ Summary finished, at 2024-07-09 08:41:11.28411
#> Washout specified is 0 meaning nothing will be applied. Suggest rechecking the
#> priorUseWashout.
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-09 08:41:28.65661
#> ✔ Summary finished, at 2024-07-09 08:41:28.694596
# }