Generates a cohort of the drug use of a certain list of concepts.

Generates a cohort of the drug use of a certain list of concepts.

Usage

generateDrugUtilisationCohortSet(
  cdm,
  name,
  conceptSetList,
  summariseMode = "AllEras",
  fixedTime = NULL,
  daysPriorHistory = 0,
  gapEra = 30,
  priorUseWashout = 0,
  cohortDateRange = as.Date(c(NA, NA)),
  imputeDuration = "eliminate",
  durationRange = c(1, Inf)
)

Arguments

cdm

A cdm_reference object.

name

Name of the GeneratedCohortSet

conceptSetList

Named list of concept sets.

summariseMode

Choice on how to summarize the exposures. There are three options: "FixedTime" each individual is followed the exact same number of days specified in 'fixedTime' argument. "AllEras" we summarize the output will be a summary of the exposed eras of each individual. Each individual can contribute multiple times. "FirstEra" we only consider the first observable era of each individual. In this case each individual can not contribute with multiple rows. By default: "AllEras".

fixedTime

Time period after first exposure where we summarize the ingredient of interest. Argument only considered if 'summariseMode' = "FixedTime". No default value is provided.

daysPriorHistory

Minimum number of days of prior history (observation time) required for the incident eras to be considered. By default: 0, meaning it has to be in observation_period table. When Null, we do not check if in observation_period table.

gapEra

Number of days between two continuous exposures to be considered in the same era. By default: 180.

priorUseWashout

Prior days without exposure. By default: NULL.

cohortDateRange

Range for cohort_start_date and cohort_end_date

imputeDuration

Whether/how the duration should be imputed "eliminate", "median", "mean", "quantile25", "quantile75".

durationRange

Range between the duration must be comprised. It should be a numeric vector of length two, with no NAs and the first value should be equal or smaller than the second one. It is only required if imputeDuration = TRUE. If NULL no restrictions are applied. By default: NULL.

Value

The function returns the 'cdm' object with the created tables as references of the object.

Examples

# \donttest{
#' library(DrugUtilisation)
library(CodelistGenerator)
library(CDMConnector)

cdm <- mockDrugUtilisation()

druglist <- getDrugIngredientCodes(cdm, c("acetaminophen", "metformin"))

cdm <- generateDrugUtilisationCohortSet(
  cdm = cdm,
  name = "drug_cohorts",
  conceptSetList = druglist,
  daysPriorHistory = 365
)

cdm$drug_cohorts
#> # Source:   table<main.drug_cohorts> [9 x 4]
#> # Database: DuckDB 0.8.0 [unknown@Linux 5.15.0-1038-azure:R 4.3.0/:memory:]
#>   cohort_definition_id subject_id cohort_start_date cohort_end_date
#>                  <int>      <dbl> <date>            <date>         
#> 1                    1          7 2018-05-21        2018-07-20     
#> 2                    2          3 2005-06-02        2008-05-08     
#> 3                    1          5 2020-05-08        2020-05-08     
#> 4                    1         10 2004-12-17        2005-08-12     
#> 5                    1          7 2017-07-11        2017-10-04     
#> 6                    1          3 2009-04-18        2009-11-07     
#> 7                    2          3 1994-02-27        1995-08-16     
#> 8                    1          5 2017-08-29        2019-12-14     
#> 9                    1         10 2003-08-02        2004-05-06     

cohortSet(cdm$drug_cohorts)
#> # A tibble: 2 × 12
#>   cohort_definition_id cohort_name  summarise_mode fixed_time days_prior_history
#>                  <int> <chr>        <chr>          <chr>      <chr>             
#> 1                    1 Ingredient:… AllEras        NA         365               
#> 2                    2 Ingredient:… AllEras        NA         365               
#> # ℹ 7 more variables: gap_era <chr>, prior_use_washout <chr>,
#> #   cohort_date_range_start <chr>, cohort_date_range_end <chr>,
#> #   impute_duration <chr>, duration_range_min <chr>, duration_range_max <chr>

cohortCount(cdm$drug_cohorts)
#> # A tibble: 2 × 3
#>   cohort_definition_id number_records number_subjects
#>                  <int>          <dbl>           <dbl>
#> 1                    1              7               4
#> 2                    2              2               1

cohortAttrition(cdm$drug_cohorts)
#> # A tibble: 16 × 7
#>    cohort_definition_id number_records number_subjects reason_id reason         
#>                   <int>          <dbl>           <dbl>     <dbl> <chr>          
#>  1                    2              8               4         1 Qualifying ini…
#>  2                    2              8               4         2 Duration imput…
#>  3                    2              5               4         3 Join eras      
#>  4                    2              2               1         4 daysPriorHisto…
#>  5                    2              2               1         5 priorUseWashou…
#>  6                    2              2               1         6 cohort_start_d…
#>  7                    2              2               1         7 cohort_end_dat…
#>  8                    2              2               1         8 summariseMode:…
#>  9                    1             16               9         1 Qualifying ini…
#> 10                    1             16               9         2 Duration imput…
#> 11                    1             13               9         3 Join eras      
#> 12                    1              7               4         4 daysPriorHisto…
#> 13                    1              7               4         5 priorUseWashou…
#> 14                    1              7               4         6 cohort_start_d…
#> 15                    1              7               4         7 cohort_end_dat…
#> 16                    1              7               4         8 summariseMode:…
#> # ℹ 2 more variables: excluded_records <dbl>, excluded_subjects <dbl>
# }