This function is used to summarise the dose table over multiple cohorts.

This function is used to summarise the dose table over multiple cohorts.

Usage

summariseDrugUse(
  cohort,
  cdm = attr(cohort, "cdm_reference"),
  strata = list(),
  drugUseEstimates = c("min", "q05", "q25", "median", "q75", "q95", "max", "mean", "sd",
    "missing"),
  minCellCount = 5
)

Arguments

cohort

Cohort with drug use variables and strata

cdm

cdm_reference generated by CDMConnector

strata

Stratification list

drugUseEstimates

Estimates that we want for the columns

minCellCount

Below this number counts will be supressed

Value

A summary of the drug use stratified by cohort_name and strata_name

Examples

# \donttest{
library(DrugUtilisation)
library(PatientProfiles)
library(CodelistGenerator)

cdm <- mockDrugUtilisation()
cdm <- generateDrugUtilisationCohortSet(
  cdm, "dus_cohort", getDrugIngredientCodes(cdm, "acetaminophen")
)
cdm[["dus_cohort"]] <- cdm[["dus_cohort"]] %>%
  addDrugUse(cdm, 1125315)
result <- summariseDrugUse(cdm[["dus_cohort"]], cdm)
#> Warning: There was 1 warning in `dplyr::summarise()`.
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "Overall"`.
#> Caused by warning in `base::min()`:
#> ! no non-missing arguments to min; returning Inf
#> Warning: There was 1 warning in `dplyr::summarise()`.
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "Overall"`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
print(result)
#> # A tibble: 101 × 11
#>    group_name  group_level   strata_name strata_level variable    variable_level
#>    <chr>       <chr>         <chr>       <chr>        <chr>       <chr>         
#>  1 cohort_name acetaminophen Overall     Overall      number sub… NA            
#>  2 cohort_name acetaminophen Overall     Overall      number rec… NA            
#>  3 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  4 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  5 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  6 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  7 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  8 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  9 cohort_name acetaminophen Overall     Overall      duration    NA            
#> 10 cohort_name acetaminophen Overall     Overall      duration    NA            
#> # ℹ 91 more rows
#> # ℹ 5 more variables: variable_type <chr>, estimate_type <chr>, estimate <chr>,
#> #   cdm_name <chr>, result_type <chr>

cdm[["dus_cohort"]] <- cdm[["dus_cohort"]] %>%
  addSex(cdm) %>%
  addAge(cdm, ageGroup = list("<40" = c(0, 30), ">40" = c(40, 150)))

summariseDrugUse(
  cdm[["dus_cohort"]], cdm, strata = list(
   "age_group" = "age_group", "sex" = "sex",
   "age_group and sex" = c("age_group", "sex")
  )
)
#> Warning: There was 1 warning in `dplyr::summarise()`.
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "Overall"`.
#> Caused by warning in `base::min()`:
#> ! no non-missing arguments to min; returning Inf
#> Warning: There was 1 warning in `dplyr::summarise()`.
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "Overall"`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
#> Warning: There were 3 warnings in `dplyr::summarise()`.
#> The first warning was:
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "<40"`.
#> Caused by warning in `base::min()`:
#> ! no non-missing arguments to min; returning Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 2 remaining warnings.
#> Warning: There were 3 warnings in `dplyr::summarise()`.
#> The first warning was:
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "<40"`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 2 remaining warnings.
#> Warning: There were 2 warnings in `dplyr::summarise()`.
#> The first warning was:
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "Female"`.
#> Caused by warning in `base::min()`:
#> ! no non-missing arguments to min; returning Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `dplyr::summarise()`.
#> The first warning was:
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "Female"`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `dplyr::summarise()`.
#> The first warning was:
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "<40 and Female"`.
#> Caused by warning in `base::min()`:
#> ! no non-missing arguments to min; returning Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 4 warnings in `dplyr::summarise()`.
#> The first warning was:
#> ℹ In argument: `dplyr::across(...)`.
#> ℹ In group 1: `strata_level = "<40 and Female"`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> # A tibble: 1,010 × 11
#>    group_name  group_level   strata_name strata_level variable    variable_level
#>    <chr>       <chr>         <chr>       <chr>        <chr>       <chr>         
#>  1 cohort_name acetaminophen Overall     Overall      number sub… NA            
#>  2 cohort_name acetaminophen Overall     Overall      number rec… NA            
#>  3 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  4 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  5 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  6 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  7 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  8 cohort_name acetaminophen Overall     Overall      duration    NA            
#>  9 cohort_name acetaminophen Overall     Overall      duration    NA            
#> 10 cohort_name acetaminophen Overall     Overall      duration    NA            
#> # ℹ 1,000 more rows
#> # ℹ 5 more variables: variable_type <chr>, estimate_type <chr>, estimate <chr>,
#> #   cdm_name <chr>, result_type <chr>
# }