Create a table with discontinuation as survival results

Create a table with discontinuation as survival results

Usage

tableDiscontinuationAsSurvival(
  result,
  header = c("cdm_name"),
  groupColumn = c("cohort_name", strataColumns(result)),
  type = NULL,
  gapSummary = TRUE,
  hide = c("variable_level"[!gapSummary], "competing_outcome", "estimate_gap",
    "event_gap", "follow_up_days", "cohort_survival_version"),
  style = NULL,
  .options = list()
)

Arguments

result

A summarised_result object.

header

Columns to use as header. See options with availableTableColumns(result).

groupColumn

Columns to group by. See options with availableTableColumns(result).

type

Character string specifying the desired output table format. See visOmopResults::tableType() for supported table types. If type = NULL, global options (set via visOmopResults::setGlobalTableOptions()) will be used if available; otherwise, a default 'gt' table is created.

gapSummary

Whether to include Gap summary statistics.

hide

Columns to hide from the visualisation. See options with availableTableColumns(result).

style

Defines the visual formatting of the table. This argument can be provided in one of the following ways:

Pre-defined style: Use the name of a built-in style (e.g., "darwin"). See visOmopResults::tableStyle() for available options.

YAML file path: Provide the path to an existing .yml file defining a new style.

List of custome R code: Supply a block of custom R code or a named list describing styles for each table section. This code must be specific to the selected table type.

If style = NULL, the function will use global options (see visOmopResults::setGlobalTableOptions()) or an existing _brand.yml file (if found); otherwise, the default style is applied. For more details, see the Styles vignette in visOmopResults website.

.options

A named list with additional formatting options. visOmopResults::tableOptions() shows allowed arguments and their default values.

Value

A table with a formatted version of summariseDiscontinuationAsSurvival() results.

Examples

# \donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

result <- summariseDiscontinuationAsSurvival(cdm$cohort1)
#> ℹ Calculating discontinuation for cohort_1.
#> ℹ Subsetting table to cohort of interest.
#> ℹ Preparing discontinuation (outcome) cohort.
#> ℹ Estimate single event survival for cohort: cohort_1 and outcome:
#>   discontinuation_of_cohort_1.
#> - Getting survival for target cohort 'cohort_1' and outcome cohort
#> 'discontinuation_of_cohort_1'
#> Getting overall estimates
#> `eventgap`, `outcome_washout`, `censor_on_cohort_exit`, `follow_up_days`, and
#> `minimum_survival_days` casted to character.
#> ✔ Discontinuation analysis for cohort_1 completed in 1s.
#> ℹ Calculating discontinuation for cohort_2.
#> ℹ Subsetting table to cohort of interest.
#> ℹ Preparing discontinuation (outcome) cohort.
#> ℹ Estimate single event survival for cohort: cohort_2 and outcome:
#>   discontinuation_of_cohort_2.
#> - Getting survival for target cohort 'cohort_2' and outcome cohort
#> 'discontinuation_of_cohort_2'
#> Getting overall estimates
#> `eventgap`, `outcome_washout`, `censor_on_cohort_exit`, `follow_up_days`, and
#> `minimum_survival_days` casted to character.
#> ✔ Discontinuation analysis for cohort_2 completed in 1s.
#> ℹ Calculating discontinuation for cohort_3.
#> ℹ Subsetting table to cohort of interest.
#> ℹ Preparing discontinuation (outcome) cohort.
#> ℹ Estimate single event survival for cohort: cohort_3 and outcome:
#>   discontinuation_of_cohort_3.
#> - Getting survival for target cohort 'cohort_3' and outcome cohort
#> 'discontinuation_of_cohort_3'
#> Getting overall estimates
#> `eventgap`, `outcome_washout`, `censor_on_cohort_exit`, `follow_up_days`, and
#> `minimum_survival_days` casted to character.
#> ✔ Discontinuation analysis for cohort_3 completed in 1s.

tableDiscontinuationAsSurvival(result)
#> cdm_name, cohort_name, cohort_survival_version, competing_outcome,
#> estimate_gap, event_gap, and follow_up_days are missing in `columnOrder`, will
#> be added last.

  

Variable name
      	Time (days)
      	Estimate name
      	
        
Data source

      
    

DUS MOCK
    

cohort_1
    

Summary statistics of discontinuation_of_cohort_1 (Outcome)
	–
	Number records
	2

	
	N events
	2

	
	Restricted mean survival (95% CI)
	417.00 (86.00, 748.00)

	
	Median survival (95% CI)
	–

	
	0% quantile (95% CI)
	0.00 (0.00, 0.00)

	
	5% quantile (95% CI)
	–

	
	25% quantile (95% CI)
	–

	
	75% quantile (95% CI)
	–

	
	95% quantile (95% CI)
	–

	
	100% quantile (95% CI)
	–

Gap summary of discontinuation_of_cohort_1 (Outcome)
	0
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	30
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	60
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	90
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	120
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	150
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	180
	N at risk
	1

	
	N events
	1

	
	N censor
	0

	210
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	240
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	270
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	300
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	330
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	360
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	390
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	420
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	450
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	480
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	510
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	540
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	570
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	600
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	630
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	656
	N at risk
	1

	
	N events
	1

	
	N censor
	0

cohort_2
    

Summary statistics of discontinuation_of_cohort_2 (Outcome)
	–
	Number records
	6

	
	N events
	6

	
	Restricted mean survival (95% CI)
	94.00 (15.00, 174.00)

	
	Median survival (95% CI)
	–

	
	0% quantile (95% CI)
	0.00 (0.00, 0.00)

	
	5% quantile (95% CI)
	–

	
	25% quantile (95% CI)
	–

	
	75% quantile (95% CI)
	–

	
	95% quantile (95% CI)
	–

	
	100% quantile (95% CI)
	–

Gap summary of discontinuation_of_cohort_2 (Outcome)
	0
	N at risk
	6

	
	N events
	0

	
	N censor
	0

	30
	N at risk
	4

	
	N events
	2

	
	N censor
	0

	60
	N at risk
	3

	
	N events
	1

	
	N censor
	0

	90
	N at risk
	2

	
	N events
	1

	
	N censor
	0

	120
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	150
	N at risk
	1

	
	N events
	1

	
	N censor
	0

	180
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	210
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	240
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	270
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	288
	N at risk
	1

	
	N events
	1

	
	N censor
	0

cohort_3
    

Summary statistics of discontinuation_of_cohort_3 (Outcome)
	–
	Number records
	2

	
	N events
	2

	
	Restricted mean survival (95% CI)
	741.00 (354.00, 1,128.00)

	
	Median survival (95% CI)
	–

	
	0% quantile (95% CI)
	0.00 (0.00, 0.00)

	
	5% quantile (95% CI)
	–

	
	25% quantile (95% CI)
	–

	
	75% quantile (95% CI)
	–

	
	95% quantile (95% CI)
	–

	
	100% quantile (95% CI)
	–

Gap summary of discontinuation_of_cohort_3 (Outcome)
	0
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	30
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	60
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	90
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	120
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	150
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	180
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	210
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	240
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	270
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	300
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	330
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	360
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	390
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	420
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	450
	N at risk
	2

	
	N events
	0

	
	N censor
	0

	480
	N at risk
	1

	
	N events
	1

	
	N censor
	0

	510
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	540
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	570
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	600
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	630
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	660
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	690
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	720
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	750
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	780
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	810
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	840
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	870
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	900
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	930
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	960
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	990
	N at risk
	1

	
	N events
	0

	
	N censor
	0

	1020
	N at risk
	1

	
	N events
	1

	
	N censor
	0

# }