Format a dose_coverage object into a visual table.

Usage

tableDoseCoverage(
  result,
  header = c("variable", "estimate"),
  splitStrata = TRUE,
  ingridientName = TRUE,
  cdmName = TRUE,
  groupColumn = NULL,
  type = "gt",
  formatEstimateName = c(`N (%)` = "<count_missing> (<percentage_missing> %)", N =
    "<count>", `Mean (SD)` = "<mean> (<sd>)", `Median (Q25 - Q75)` =
    "<median> (<q25> - <q75>)"),
  .options = list()
)

Arguments

result

A summarised_result object with results from summariseDoseCoverage().

header

A vector containing which elements should go into the header in order. Allowed are: cdm_name, group, strata, variable, and estimate.

splitStrata

If TRUE strata columns will be splitted.

ingridientName

If TRUE cohort names will be displayed.

cdmName

If TRUE database names will be displayed.

groupColumn

Column to use as group labels.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

.options

Named list with additional formatting options. DrugUtilisation::optionstableDoseCoverage() shows allowed arguments and their default values.

Value

A table with a formatted version of summariseDrugCoverage() results.

Examples

# \donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

result <- summariseDoseCoverage(cdm, 1125315)
#> ℹ The following estimates will be computed:
#> • daily_dose: count_missing, percentage_missing, mean, sd, q25, median, q75
#> ! Table is collected to memory as not all requested estimates are supported on
#>   the database side
#> → Start summary of data, at 2024-07-08 09:36:12.812282
#> ✔ Summary finished, at 2024-07-08 09:36:13.088028

tableDoseCoverage(result)

  

Database name
      	Ingredient name
      	Unit
      	Route
      	Pattern id
      	
        Number records
      
      	
        Missing dose
      
      	
        Daily dose
      
    

N
      	N (%)
      	Mean (SD)
      	Median (Q25 - Q75)
    

DUS MOCK
	Acetaminophen
	Overall
	Overall
	Overall
	13
	0 (0.00 %)
	15,354.72 (40,814.75)
	233.33 (61.27 - 2,128.60)

	
	Milligram
	Overall
	Overall
	13
	0 (0.00 %)
	15,354.72 (40,814.75)
	233.33 (61.27 - 2,128.60)

	
	
	Oral
	Overall
	4
	0 (0.00 %)
	743.94 (1,285.41)
	152.22 (54.50 - 841.67)

	
	
	Topical
	Overall
	9
	0 (0.00 %)
	21,848.40 (48,414.57)
	573.77 (61.27 - 2,128.60)

	
	
	Oral
	9
	4
	0 (0.00 %)
	743.94 (1,285.41)
	152.22 (54.50 - 841.67)

	
	
	Topical
	18
	3
	0 (0.00 %)
	64,709.53 (72,396.68)
	48,000.00 (25,064.30 - 96,000.00)

	
	
	
	9
	6
	0 (0.00 %)
	417.83 (686.46)
	73.56 (35.64 - 451.79)

CDMConnector::cdmDisconnect(cdm = cdm)
# }