Format a drug_utilisation object into a visual table.

Usage

tableDrugUtilisation(
  result,
  header = c("group", "strata"),
  splitStrata = TRUE,
  cohortName = TRUE,
  cdmName = TRUE,
  conceptSet = TRUE,
  ingredient = TRUE,
  groupColumn = NULL,
  type = "gt",
  formatEstimateName = c(`N (%)` = "<count_missing> (<percentage_missing> %)", N =
    "<count>", `Mean (SD)` = "<mean> (<sd>)", `Median (Q25 - Q75)` =
    "<median> (<q25> - <q75>)"),
  .options = list()
)

Arguments

result

A summarised_result object with results from summariseDrugUtilisation().

header

A vector containing which elements should go into the header in order. Allowed are: cdm_name, group, strata, variable, estimate.

splitStrata

If TRUE strata columns will be split.

cohortName

If TRUE cohort names will be displayed.

cdmName

If TRUE database names will be displayed.

conceptSet

If TRUE concept sets name will be displayed.

ingredient

If TRUE ingredients names will be displayed for dose calculation.

groupColumn

Column to use as group labels, these can be: "cdm_name", "cohort_name", "concept_set", "variable_name", and/or "ingredient". If strata is split, any of the levels can be used, otherwise "strata_name" and "strata_level" can be used for table group format.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

.options

Named list with additional formatting options. DrugUtilisation::defaultTableOptions() shows allowed arguments and their default values.

Value

A table with a formatted version of summariseIndication() results.

Examples

# \donttest{

library(DrugUtilisation)
library(CodelistGenerator)

cdm <- mockDrugUtilisation()
cdm <- generateDrugUtilisationCohortSet(
  cdm, "dus_cohort", getDrugIngredientCodes(cdm, name = "acetaminophen")
)
cdm[["dus_cohort"]] %>%
  summariseDrugUtilisation(ingredientConceptId = 1125315) |>
  tableDrugUtilisation()
#> ! Results have not been suppressed.

  

Database name
      	Variable
      	Unit
      	Estimate name
      	Concept set
      	Ingredient
      	
        Cohort name
      
    

161 acetaminophen
    

DUS MOCK
	Number records
	-
	N
	Overall
	Overall
	11

	Number subjects
	-
	N
	Overall
	Overall
	9

	Number exposures
	-
	N (%)
	Ingredient 1125315 descendants
	Overall
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Overall
	1.18 (0.60)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Overall
	1 (1 - 1)

	Time to exposure
	-
	N (%)
	Ingredient 1125315 descendants
	Overall
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Overall
	0.00 (0.00)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Overall
	0 (0 - 0)

	Cumulative quantity
	-
	N (%)
	Ingredient 1125315 descendants
	Overall
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Overall
	50.55 (33.26)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Overall
	40.00 (35.00 - 57.50)

	Initial quantity
	-
	N (%)
	Ingredient 1125315 descendants
	Overall
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Overall
	41.36 (24.71)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Overall
	35.00 (30.00 - 42.50)

	Number eras
	-
	N (%)
	Ingredient 1125315 descendants
	Overall
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Overall
	1.00 (0.00)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Overall
	1 (1 - 1)

	Exposed time
	-
	N (%)
	Ingredient 1125315 descendants
	Overall
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Overall
	306.09 (571.79)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Overall
	61 (12 - 229)

	Cumulative dose
	Milligram
	N (%)
	Ingredient 1125315 descendants
	Acetaminophen
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Acetaminophen
	151,900.00 (298,018.82)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Acetaminophen
	20,000.00 (16,750.00 - 68,000.00)

	Initial daily dose
	Milligram
	N (%)
	Ingredient 1125315 descendants
	Acetaminophen
	0 (0.00 %)

	
	
	Mean (SD)
	Ingredient 1125315 descendants
	Acetaminophen
	17,952.55 (44,164.85)

	
	
	Median (Q25 - Q75)
	Ingredient 1125315 descendants
	Acetaminophen
	233.33 (66.19 - 2,208.33)

# }