Create a table with proportion of patients covered results

Usage

tableProportionOfPatientsCovered(
  result,
  times = NULL,
  header = c("group", "strata"),
  splitStrata = TRUE,
  cohortName = TRUE,
  cdmName = TRUE,
  groupColumn = "variable_name",
  type = "gt",
  .options = list()
)

Arguments

result

A summarised_result object with results from summariseProportionOfPatientsCovered().

times

Days to include in the table. If NULL all days will be included.

header

A vector containing which elements should go into the header in order. Allowed are: cdm_name, group, strata, variable.

splitStrata

If TRUE strata columns will be split.

cohortName

If TRUE cohort names will be displayed.

cdmName

If TRUE database names will be displayed.

groupColumn

Column to use as group labels.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

.options

Named list with additional formatting options. DrugUtilisation::defaultTableOptions() shows allowed arguments and their default values.

Value

A table with a formatted version of summariseProportionOfPatientsCovered() results.