DrugUtilisation: Summarise Patient-Level Drug Utilisation in Data Mapped to the OMOP Common Data Model

Summarise patient-level drug utilisation cohorts using data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model. New users and prevalent users cohorts can be generated and their characteristics, indication and drug use summarised.

See also

Author

Maintainer: Martí Català marti.catalasabate@ndorms.ox.ac.uk (ORCID)

Authors:

Mike Du mike.du@ndorms.ox.ac.uk (ORCID)

Yuchen Guo yuchen.guo@ndorms.ox.ac.uk (ORCID)

Kim Lopez-Guell kim.lopez@spc.ox.ac.uk (ORCID)

Nuria Mercade-Besora nuria.mercadebesora@ndorms.ox.ac.uk (ORCID)

Xihang Chen xihang.chen@ndorms.ox.ac.uk (ORCID)

Other contributors:

Edward Burn edward.burn@ndorms.ox.ac.uk (ORCID) [contributor]

Xintong Li xintong.li@ndorms.ox.ac.uk (ORCID) [contributor]

Marta Alcalde-Herraiz marta.alcaldeherraiz@ndorms.ox.ac.uk (ORCID) [contributor]