Add new columns with dose related information to the cohort

Add new columns with dose related information to the cohort

Usage

addDrugUse(
  cohort,
  cdm,
  ingredientConceptId,
  conceptSetList = NULL,
  initialDailyDose = TRUE,
  numberExposures = TRUE,
  duration = TRUE,
  cumulativeDose = TRUE,
  numberEras = TRUE,
  initialQuantity = TRUE,
  cumulativeQuantity = TRUE,
  gapEra = 30,
  eraJoinMode = "Zero",
  overlapMode = "Sum",
  sameIndexMode = "Sum",
  imputeDuration = "eliminate",
  imputeDailyDose = "eliminate",
  durationRange = c(1, Inf),
  dailyDoseRange = c(0, Inf)
)

Arguments

cohort

Cohort in the cdm

cdm

cdm_reference created with CDMConnector::cdmFromCon

ingredientConceptId

Ingredient OMOP concept that we are interested for the study. It is a compulsory input, no default value is provided.

conceptSetList

List of concepts to be included. If NULL all the descendants of ingredient concept id will be used. By default: NULL.

initialDailyDose

Whether initial dose should be displayed

numberExposures

Whether number exposures should be displayed

duration

Whether duration should be displayed

cumulativeDose

Whether cumulative dose should be displayed

numberEras

Whether the number of eras should be displayed

initialQuantity

Whether initial quantity should be displayed

cumulativeQuantity

Whether cumulative quantity should be displayed

gapEra

Number of days between two continuous exposures to be considered in the same era. By default: 180.

eraJoinMode

How two different continuous exposures are joined in an era. There are four options: "Zero" the exposures are joined considering that the period between both continuous exposures the subject is treated with a daily dose of zero. The time between both exposures contributes to the total exposed time. "Join" the exposures are joined considering that the period between both continuous exposures the subject is treated with a daily dose of zero. The time between both exposures does not contribute to the total exposed time. "Previous" the exposures are joined considering that the period between both continuous exposures the subject is treated with the daily dose of the previous subexposure. The time between both exposures contributes to the total exposed time. "Subsequent" the exposures are joined considering that the period between both continuous exposures the subject is treated with the daily dose of the subsequent subexposure. The time between both exposures contributes to the total exposed time. By default: "Previous".

overlapMode

How the overlapping between two exposures that do not start on the same day is solved inside a subexposure. There are five possible options: "Previous" the considered daily_dose is the one of the earliest exposure. "Subsequent" the considered daily_dose is the one of the new exposure that starts in that subexposure. "Minimum" the considered daily_dose is the minimum of all of the exposures in the subexposure. "Maximum" the considered daily_dose is the maximum of all of the exposures in the subexposure. "Sum" the considered daily_dose is the sum of all the exposures present in the subexposure. By default: "Previous".

sameIndexMode

How the overlapping between two exposures that start on the same day is solved inside a subexposure. There are five possible options: "Minimum" the considered daily_dose is the minimum of all of the exposures in the subexposure. "Maximum" the considered daily_dose is the maximum of all of the exposures in the subexposure. "Sum" the considered daily_dose is the sum of all the exposures present in the subexposure. By default: "Sum".

imputeDuration

Whether/how the duration should be imputed "eliminate", "median", "mean", "quantile25", "quantile75". . By default: eliminate

imputeDailyDose

Whether/how the daily_dose should be imputed "eliminate", "median", "mean", "quantile25", "quantile75". By default: "eliminate"

durationRange

Range between the duration must be comprised. It should be a numeric vector of length two, with no NAs and the first value should be equal or smaller than the second one. It is only required if imputeDuration = TRUE. If NULL no restrictions are applied. By default: NULL.

dailyDoseRange

Range between the daily_dose must be comprised. It should be a numeric vector of length two, with no NAs and the first value should be equal or smaller than the second one. It is only required if imputeDailyDose = TRUE. If NULL no restrictions are applied. By default: NULL.

Value

The function returns the dose information for all the individuals of dusCohortName.

Examples

# \donttest{
library(DrugUtilisation)
library(CodelistGenerator)

cdm <- mockDrugUtilisation()
cdm <- generateDrugUtilisationCohortSet(
  cdm, "dus_cohort", getDrugIngredientCodes(cdm, "acetaminophen")
)
cdm$dus_cohort %>%
  addDrugUse(cdm, 1125315)
#> Warning: Different names in conceptSetList (acetaminophen) than in the created
#> cohortSet.
#> Warning: NAs introduced by coercion
#> # Source:   table<dbplyr_046> [?? x 11]
#> # Database: DuckDB 0.8.1 [unknown@Linux 5.15.0-1041-azure:R 4.3.1/:memory:]
#>    cohort_definition_id subject_id cohort_start_date cohort_end_date duration
#>                   <int>      <dbl> <date>            <date>             <dbl>
#>  1                    1          7 2017-07-11        2017-10-04            86
#>  2                    1          7 2018-05-21        2018-07-20            61
#>  3                    1          5 2020-05-08        2020-05-08             1
#>  4                    1         10 2004-12-17        2005-08-12           239
#>  5                    1          2 2022-06-06        2022-06-07             2
#>  6                    1          1 2021-02-25        2021-03-11            15
#>  7                    1          7 2016-02-01        2017-02-03           369
#>  8                    1          3 2009-04-18        2009-11-07           204
#>  9                    1          6 1990-03-07        1991-11-24           628
#> 10                    1          4 2021-05-05        2021-12-15           225
#> # ℹ more rows
#> # ℹ 6 more variables: number_exposures <dbl>, cumulative_quantity <dbl>,
#> #   initial_quantity <dbl>, initial_daily_dose <dbl>, number_eras <dbl>,
#> #   cumulative_dose <dbl>
# }