Run benchmark of drug utilisation cohort generation

Run benchmark of drug utilisation cohort generation

Usage

benchmarkDUS(
  cdm,
  numberOfCohort = 1:4,
  indicationCohortName = "cohort1",
  ingredientId = 1125315,
  drugExposureName = "drug_exposure"
)

Arguments

cdm

A CDM reference object

numberOfCohort

Number of cohort to generate for benchmarking. An integer or a vector of integers

indicationCohortName

Name of indication cohort table

ingredientId

Ingredient OMOP concept that we are interested for the study. It is a compulsory input, no default value is provided.

drugExposureName

Name of drug_exposure table in cdm, the table must contain drug_concept_id, quantity, drug_exposure_start_date and drug_exposure_end_date as columns

Value

a tibble with time taken for different analyses

Examples

# \donttest{
cdm <- DrugUtilisation::mockDrugUtilisation()
timings <- DrugUtilisation::benchmarkDUS(cdm)
#> Warning: The `cdm` argument of `summariseDrugUse()` is deprecated as of DrugUtilisation
#> 0.5.0.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
# }