Run benchmark of drug utilisation cohort generation

Run benchmark of drug utilisation cohort generation

Usage

benchmarkDrugUtilisation(
  cdm,
  numberOfCohort = 1:4,
  indicationCohortName = "cohort1",
  ingredientId = 1125315,
  drugExposureName = "drug_exposure"
)

Arguments

cdm

A CDM reference object

numberOfCohort

Number of cohort to generate for benchmarking. An integer or a vector of integers

indicationCohortName

Name of indication cohort table

ingredientId

Ingredient OMOP concept that we are interested for the study. It is a compulsory input, no default value is provided.

drugExposureName

Name of drug_exposure table in cdm, the table must contain drug_concept_id, quantity, drug_exposure_start_date and drug_exposure_end_date as columns

Value

a tibble with time taken for different analyses

Examples

# \donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

timings <- benchmarkDrugUtilisation(cdm)
#> Warning: The `durationRange` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.7.0.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `imputeDuration` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.7.0.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `priorUseWashout` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.7.0.
#> ℹ Please use `requirePriorDrugWashout()` instead.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `priorObservation` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.7.0.
#> ℹ Please use `requireObservationBeforeDrug()` instead.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `cohortDateRange` argument of `generateDrugUtilisationCohortSet()` is
#> deprecated as of DrugUtilisation 0.7.0.
#> ℹ Please use `requireDrugInDateRange()` instead.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `limit` argument of `generateDrugUtilisationCohortSet()` is deprecated as
#> of DrugUtilisation 0.7.0.
#> ℹ Please use `requireIsFirstDrugEntry()` instead.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> Warning: The `cdm` argument of `summariseDrugUse()` is deprecated as of DrugUtilisation
#> 0.5.0.
#> ℹ The deprecated feature was likely used in the DrugUtilisation package.
#>   Please report the issue to the authors.
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-18 14:41:59.513069
#> ✔ Summary finished, at 2024-07-18 14:41:59.553784
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-18 14:42:17.016478
#> ✔ Summary finished, at 2024-07-18 14:42:17.052974
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-18 14:42:34.627105
#> ✔ Summary finished, at 2024-07-18 14:42:34.663564
#> ! names of group will be ignored
#> ℹ The following estimates will be computed:
#> → Start summary of data, at 2024-07-18 14:42:52.140564
#> ✔ Summary finished, at 2024-07-18 14:42:52.178379

timings
#> # A tibble: 17 × 5
#>    task                time_taken_secs time_taken_mins time_taken_hours person_n
#>    <chr>                         <dbl>           <dbl>            <dbl>    <dbl>
#>  1 DUS 1 cohorts                  3.21            0.05                0       10
#>  2 add indication 1 c…            5.02            0.08                0       10
#>  3 add drug use for 1…            8.39            0.14                0       10
#>  4 summarise drug use…            8.39            0.14                0       10
#>  5 DUS 2 cohorts                  3.19            0.05                0       10
#>  6 add indication 2 c…            5               0.08                0       10
#>  7 add drug use for 2…            8.37            0.14                0       10
#>  8 summarise drug use…            8.37            0.14                0       10
#>  9 DUS 3 cohorts                  3.11            0.05                0       10
#> 10 add indication 3 c…            5.12            0.09                0       10
#> 11 add drug use for 3…            8.45            0.14                0       10
#> 12 summarise drug use…            8.45            0.14                0       10
#> 13 DUS 4 cohorts                  3.1             0.05                0       10
#> 14 add indication 4 c…            5.16            0.09                0       10
#> 15 add drug use for 4…            8.35            0.14                0       10
#> 16 summarise drug use…            8.35            0.14                0       10
#> 17 add daily dose                 0.99            0.02                0       10
# }