Generates a cohort of the drug use of ATC name(s) of interest.

Generates a cohort of the drug use of ATC name(s) of interest.

Usage

generateAtcCohortSet(
  cdm,
  name,
  atcName = NULL,
  level = c("ATC 1st"),
  doseForm = NULL,
  gapEra = 1,
  durationRange = lifecycle::deprecated(),
  imputeDuration = lifecycle::deprecated(),
  priorUseWashout = lifecycle::deprecated(),
  priorObservation = lifecycle::deprecated(),
  cohortDateRange = lifecycle::deprecated(),
  limit = lifecycle::deprecated()
)

Arguments

cdm

A cdm_reference object.

name

Name of the GeneratedCohortSet

atcName

Names of ATC of interest.

level

ATC level. Can be one or more of "ATC 1st", "ATC 2nd", "ATC 3rd", "ATC 4th", and "ATC 5th"

doseForm

Only descendants codes with the specified dose form will be returned. If NULL, descendant codes will be returned regardless of dose form.

gapEra

Number of days between two continuous exposures to be considered in the same era.

durationRange

Deprecated.

imputeDuration

Deprecated.

priorUseWashout

Deprecated

priorObservation

Deprecated.

cohortDateRange

Deprecated.

limit

Deprecated.

Value

The function returns the 'cdm' object with the created cohorts as references of the object.

Examples

# \donttest{
library(DrugUtilisation)
cdm <- mockDrugUtilisation()
cdm <- generateAtcCohortSet(cdm, name =  "test", priorObservation = 365)
#> Warning: The `priorObservation` argument of `generateAtcCohortSet()` is deprecated as of
#> DrugUtilisation 0.7.0.
#> ℹ Please use `requireObservationBeforeDrug()` instead.
#> Warning: cohort_name must be snake case, the following cohorts will be renamed:
#> • A_alimentary_tract_and_metabolism -> a_alimentary_tract_and_metabolism
cdm
#> 
#> ── # OMOP CDM reference (duckdb) of DUS MOCK ───────────────────────────────────
#> • omop tables: person, observation_period, concept, concept_ancestor,
#> drug_strength, drug_exposure, condition_occurrence, visit_occurrence,
#> observation, concept_relationship
#> • cohort tables: cohort1, cohort2, test
#> • achilles tables: -
#> • other tables: -
cdm$test
#> # Source:   table<main.test> [0 x 4]
#> # Database: DuckDB v1.0.0 [unknown@Linux 6.5.0-1023-azure:R 4.4.1/:memory:]
#> # ℹ 4 variables: cohort_definition_id <int>, subject_id <int>,
#> #   cohort_start_date <date>, cohort_end_date <date>
settings(cdm$test)
#> # A tibble: 1 × 4
#>   cohort_definition_id cohort_name                       gap_era dose_form
#>                  <int> <chr>                             <chr>   <chr>    
#> 1                    1 a_alimentary_tract_and_metabolism 1       ""       
# }