Generate a set of drug cohorts based on given concepts

Adds a new cohort table to the cdm reference with individuals who have drug exposure records with the specified concepts. Cohort start and end dates will be based on drug record start and end dates, respectively. Records that overlap or have fewer days between them than the specified gap era will be concatenated into a single cohort entry.

Usage

generateDrugUtilisationCohortSet(
  cdm,
  name,
  conceptSet,
  gapEra = 1,
  subsetCohort = NULL,
  subsetCohortId = NULL,
  numberExposures = FALSE,
  daysPrescribed = FALSE
)

Arguments

cdm

A cdm_reference object.

name

Name of the new cohort table, it must be a length 1 character vector.

conceptSet

List of concepts to be included.

gapEra

Number of days between two continuous exposures to be considered in the same era.

subsetCohort

Cohort table to subset.

subsetCohortId

Cohort id to subset.

numberExposures

Whether to include 'number_exposures' (number of drug exposure records between indexDate and censorDate).

daysPrescribed

Whether to include 'days_prescribed' (number of days prescribed used to create each era).

Value

The function returns the cdm reference provided with the addition of the new cohort table.

Examples

# \donttest{
cdm <- mockDrugUtilisation()

druglist <- CodelistGenerator::getDrugIngredientCodes(
  cdm, c("acetaminophen", "metformin"), nameStyle = "{concept_name}"
)

cdm <- generateDrugUtilisationCohortSet(
  cdm = cdm,
  name = "drug_cohorts",
  conceptSet = druglist,
  gapEra = 30,
  numberExposures = TRUE,
  daysPrescribed = TRUE
)
#> ℹ Subsetting drug_exposure table
#> ℹ Checking whether any record needs to be dropped.
#> ℹ Collapsing overlaping records.
#> ℹ Collapsing records with gapEra = 30 days.

cdm$drug_cohorts |>
  dplyr::glimpse()
#> Rows: ??
#> Columns: 6
#> Database: DuckDB v1.2.2 [unknown@Linux 6.11.0-1012-azure:R 4.5.0/:memory:]
#> $ cohort_definition_id <int> 2, 2, 2, 2, 1, 2, 2, 1, 1, 2, 1, 2, 1, 1, 2, 2, 1
#> $ subject_id           <int> 2, 2, 5, 5, 1, 10, 6, 8, 10, 10, 6, 9, 2, 4, 9, 3…
#> $ cohort_start_date    <date> 1972-08-22, 1997-02-18, 1979-12-19, 1980-08-30, 2…
#> $ cohort_end_date      <date> 1986-10-19, 2006-04-04, 1980-02-15, 1982-07-31, …
#> $ number_exposures     <int> 1, 1, 1, 1, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> $ days_prescribed      <int> 5172, 3333, 59, 701, 39, 497, 645, 1225, 2258, 8…
# }