This function is used to summarise the indication table over multiple cohorts.

This function is used to summarise the indication table over multiple cohorts.

Usage

summariseIndication(
  cohort,
  cdm = lifecycle::deprecated(),
  strata = list(),
  minCellCount = lifecycle::deprecated()
)

Arguments

cohort

Cohort with indications and strata.

cdm

Deprecated.

strata

Stratification list.

minCellCount

Deprecated.

Value

A Tibble with 4 columns: cohort_definition_id, variable, estimate and value. There will be one row for each cohort, variable and cohort combination.

Examples

# \donttest{
library(DrugUtilisation)
library(PatientProfiles)
library(CodelistGenerator)
library(CDMConnector)

cdm <- mockDrugUtilisation()
indications <- list("headache" = 378253, "asthma" = 317009)
cdm <- generateConceptCohortSet(cdm, indications, "indication_cohorts")
acetaminophen <- getDrugIngredientCodes(cdm, "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(cdm, "drug_cohort", acetaminophen)
cdm[["drug_cohort"]] <- cdm[["drug_cohort"]] %>%
  addIndication(
    indicationCohortName = "indication_cohorts",
    indicationWindow = list(c(0,0),c(-30,0),c(-365,0))
  )

# }