Summarise the indications of individuals in a drug cohort

Summarise the observed indications of patients in a drug cohort based on their presence in an indication cohort in a specified time window. If an individual is not in one of the indication cohorts, they will be considered to have an unknown indication if they are present in one of the specified OMOP CDM clinical tables. Otherwise, if they are neither in an indication cohort or a clinical table they will be considered as having no observed indication.

Usage

summariseIndication(
  cohort,
  strata = list(),
  indicationCohortName,
  indicationCohortId = NULL,
  indicationWindow = list(c(0, 0)),
  unknownIndicationTable = NULL,
  indexDate = "cohort_start_date"
)

Arguments

cohort

A cohort table in a cdm reference.

strata

List of variables to stratify results by. These variables must be present in the cohort table.

indicationCohortName

Name of the cohort table with potential indications.

indicationCohortId

The target cohort ID to add indication. If NULL all cohorts will be considered.

indicationWindow

The time window over which to identify indications.

unknownIndicationTable

Tables in the OMOP CDM to search for unknown indications.

indexDate

A date variable in the cohort table for which indications will be found relative to.

Value

A summarised result

Examples

# \donttest{
library(DrugUtilisation)
library(PatientProfiles)
library(CDMConnector)
library(dplyr)

cdm <- mockDrugUtilisation()

indications <- list("headache" = 378253, "asthma" = 317009)
cdm <- generateConceptCohortSet(cdm, indications, "indication_cohorts")

cdm <- generateIngredientCohortSet(cdm = cdm, name = "drug_cohort",
                                   ingredient = "acetaminophen")

cdm$drug_cohort |>
summariseIndication(indicationCohortName = "indication_cohorts",
                    unknownIndicationTable = "condition_occurrence",
                    indicationWindow = list(c(-Inf, 0))) |>
 glimpse()
#> ℹ The following estimates will be computed:
#> • indication_minf_to_0: count, percentage
#> → Start summary of data, at 2024-07-24 10:28:43.365731
#> ✔ Summary finished, at 2024-07-24 10:28:43.428989
#> Rows: 12
#> Columns: 13
#> $ result_id        <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
#> $ cdm_name         <chr> "DUS MOCK", "DUS MOCK", "DUS MOCK", "DUS MOCK", "DUS …
#> $ group_name       <chr> "cohort_name", "cohort_name", "cohort_name", "cohort_…
#> $ group_level      <chr> "161_acetaminophen", "161_acetaminophen", "161_acetam…
#> $ strata_name      <chr> "overall", "overall", "overall", "overall", "overall"…
#> $ strata_level     <chr> "overall", "overall", "overall", "overall", "overall"…
#> $ variable_name    <chr> "number records", "number subjects", "Indication any …
#> $ variable_level   <chr> NA, NA, "Asthma", "Asthma and Headache", "Headache", …
#> $ estimate_name    <chr> "count", "count", "count", "count", "count", "count",…
#> $ estimate_type    <chr> "integer", "integer", "integer", "integer", "integer"…
#> $ estimate_value   <chr> "11", "9", "1", "1", "1", "5", "3", "9.09090909090909…
#> $ additional_name  <chr> "overall", "overall", "overall", "overall", "overall"…
#> $ additional_level <chr> "overall", "overall", "overall", "overall", "overall"…

# }